fix(frontend+backend): fix data display, API wiring, and auth issues
FRONTEND:
- fix(vite): restructure vite.config.js — alias was at top level instead of
  inside resolve{}; build.outDir was nested inside resolve instead of top level;
  update outDir path to correct posBackend static resources location

- fix(api): correct all sale endpoints — viewSale/deleteSale used wrong HTTP
  methods and query params instead of path params; all sale calls now use
  /api/sale/* consistently; getAllSales now paginated (page, size, sort params)

- fix(transactions): replace txn.saleDate with txn.createdAt (field renamed in
  backend); implement real sale item table in expanded rows (product name, qty,
  price, line total); add retry button; add real pagination (prev/next, page info);
  disable Return button for already-returned sales; handle empty state

- fix(checkout): align createSale payload with updated SaleRequest schema —
  remove server-computed fields (totalAmount, subtotal, discountAmount, taxAmount);
  send only {cashierId, paymentMethod, items:[{productId,quantity}]}; use
  savedSale.totalAmount from response for receipt display

- fix(inventory): add keycloak.authenticated guard to useEffect to prevent 401
  on API call before token is available; add keycloak to dependency array

- fix(keycloak): remove invalid redirectUri with wildcard from initOptions;
  add pkceMethod:S256 for secure PKCE flow; redirectUri handled per-call

- fix(env): add missing VITE_API_BASE_URL to .env; fix VITE_REDIRECT_URI
  (remove /*); create .env.example template for local dev setup

BACKEND:
- fix(sales): add @NamedEntityGraph(Sales.withItems) to Sales entity to eagerly
  load saleItems and their products in list/get queries; prevents lazy
  initialization issues when serializing to JSON

- fix(sales-repo): apply @EntityGraph on findAll(Pageable) and findById(Long)
  so sale items are always included in responses; change ID type to Long

- fix(sales): update SalesService and SalesController to use Long IDs throughout

+++ posBackend/src/main/java/com/pos/service/impl/SalesService.java
@@ -77,7 +77,7 @@ public class SalesService {
     }
 
     @Transactional(readOnly = true)
-    public Sales viewSale(int id) {
+    public Sales viewSale(Long id) {
         return repository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Sale not found"));
     }
@@ -88,7 +88,7 @@ public class SalesService {
     }
 
     @Transactional
-    public void deleteSale(int id) {
+    public void deleteSale(Long id) {
         if (!repository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Sale not found");
         }
@@ -96,7 +96,7 @@ public class SalesService {
     }
 
     @Transactional
-    public Sales processReturn(int saleId, SaleReturnRequest request) {
+    public Sales processReturn(Long saleId, SaleReturnRequest request) {
         Sales sale = repository.findById(saleId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Sale not found"));
